backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.2
